requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
djangorestframework==3.9.2
gunicorn==19.9.0
